+++ test_iptc_pmd_generic_all1.py
@@ -17,6 +17,8 @@ def semiptc2seret():
     print('========== semantic IPTC TO IPTC serialized ET md ============')
     iptcmd_out = IptcPhotometadataGeneric()
     iptcmd_out.currentdir = currentdir
+    # line below: switch between the role of the deprecated IPTC location properties (IPTC Core): CREATED or SHOWN
+    iptcmd_out.deprlocationrole = IptcDeprecatedLocationRole.SHOWN
     # set properties - it follows the sequence of properties in the project sheet
     # Core schema metadata
     iptcmd_out.copyrightNotice = '(c) 2020 Copyright Pizzashots, all rights reserved'
@@ -44,7 +46,6 @@ def semiptc2seret():
     creatorContactInfo.weburlwork = 'https://janem.example.com'
     creatorExt2.creatorContactInfo = creatorContactInfo
     iptcmd_out.creatorsExt.append(creatorExt2)
-
     iptcmd_out.creditLine = "J Boltenheimer/Pizzashots"
     iptcmd_out.dateCreated = '2020-07-07T12:13:14+02:00'
     iptcmd_out.captionWriter = 'Bobby Captionwriter'
@@ -58,9 +59,17 @@ def semiptc2seret():
     iptcmd_out.sceneCodes = ['SceneCode 01', 'SceneCode 02']
     iptcmd_out.source = 'The best Source of all'
     iptcmd_out.subjectCodes = ['SubjectCode 01', 'SubjectCode 02']
-
     iptcmd_out.title = 'A Title'
     # Extension schema metadata
+    t_cvterm1 = CvTerm()
+    t_cvterm1.cvId = 'https://example.com/cvid/124123'
+    t_cvterm1.cvTermId = 'https://example.com/termid/term22'
+    t_cvterm1.cvTermName = 'Name of Term22'
+    t_cvterm2 = CvTerm()
+    t_cvterm2.cvId = 'https://example.com/cvid/58430'
+    t_cvterm2.cvTermId = 'https://example.com/termid/term55'
+    t_cvterm2.cvTermName = 'Name of Term55'
+    iptcmd_out.aboutCvTerms = [t_cvterm1, t_cvterm2]
     t_licensor = Licensor()
     t_licensor.licensorID = 'https://pizzashort.example.com'
     t_licensor.licensorName = 'Pizzashots Inc.'
@@ -82,11 +91,25 @@ def semiptc2seret():
     t_location2.countryName = 'Country A2'
     t_location2.countryCode = 'CA2'
     t_location2.worldRegion = 'Worldregion A2'
-    iptcmd_out.locationCreated = t_location1
+    iptcmd_out.locationCreated = t_location2
     iptcmd_out.locationsShown = [t_location1, t_location2]
+    t_personShown1 = PersonWDetails()
+    t_personShown1.identifiers = ['https://persons.example.com/id/2143312',
+                                  'https://otherpersons.example.com/id/2143312']
+    t_personShown1.name = 'Laura Hillinger'
+    t_personShown1.description = 'Nice photo metadata editor and a secret artist'
+    t_personShown1.characteristics = [t_cvterm1, t_cvterm2]
+    t_personShown2 = PersonWDetails()
+    t_personShown2.identifiers = ['https://persons.example.com/id/4543543',
+                                  'https://otherpersons.example.com/id/758758']
+    t_personShown2.name = 'Frank Muller'
+    t_personShown2.description = 'Reliable event oraganizer'
+    t_personShown2.characteristics = [t_cvterm2, t_cvterm1]
+    iptcmd_out.personsShown = [t_personShown1, t_personShown2]
     iptcmd_out.webstatementRights = 'https://pizzashort.example.com/copyright-licensing'
     # now finalize and embed
     iptcmd_out.export_semiptc_as_jsonfile('./semiptc-all_sem.json')
+    iptcmd_out.export_self_as_jsonfile('./iptcmd_out-self.json')
     iptcmd_out.transform_semiptc_metadata_to_seret()
     iptcmd_out.export_seret_as_jsonfile('./semiptc-all_seret.json')
     # et.etdata = iptcmd_out.seret_metadata
@@ -104,33 +127,43 @@ def seret2semiptc():
         res_log.write('Results from reading a file with serialized IPTC PMD using ExifTool naming:\n')
         # get properties
         res_log.write('\n***** IPTC Core Schema')
-        res_log.write('\nCopyright Notice: ' + iptcmd_in.copyrightNotice)
+        res_log.write('\n* Copyright Notice: ' + iptcmd_in.copyrightNotice)
         creatorsExt_list = []
         for creatorExt in iptcmd_in.creatorsExt:
             creatorExt_dict = creatorExt.todict()
             creatorsExt_list.append(creatorExt_dict)
-        res_log.write('\nCreator Extensive: ' + json.dumps(creatorsExt_list, indent=2))
-        res_log.write('\nCredit Line: ' + iptcmd_in.creditLine)
-        res_log.write('\nDate Created: ' + iptcmd_in.dateCreated)
-        res_log.write('\nCaption Writer: ' + iptcmd_in.captionWriter)
-        res_log.write('\nDescription: ' + iptcmd_in.description)
-        res_log.write('\nHeadline: ' + iptcmd_in.headline)
-        res_log.write('\nInstructions: ' + iptcmd_in.instructions)
-        res_log.write('\nIntellectual Genre: ' + iptcmd_in.intellectualGenre)
-        res_log.write('\nJob Id: ' + iptcmd_in.jobid)
-        res_log.write('\nKeywords (multiple): ' + ', '.join(iptcmd_in.keywords))
-        res_log.write('\nRights Usage Terms: ' + iptcmd_in.usageTerms)
-        res_log.write('\nScene Codes (multiple): ' + ', '.join(iptcmd_in.sceneCodes))
-        res_log.write('\nSource: ' + iptcmd_in.source)
-        res_log.write('\nSubject Codes (multiple): ' + ', '.join(iptcmd_in.subjectCodes))
-        res_log.write('\nTitle: ' + iptcmd_in.title)
+        res_log.write('\n* Creator Extensive: ' + json.dumps(creatorsExt_list, indent=2))
+        res_log.write('\n* Credit Line: ' + iptcmd_in.creditLine)
+        res_log.write('\n* Date Created: ' + iptcmd_in.dateCreated)
+        res_log.write('\n* Caption Writer: ' + iptcmd_in.captionWriter)
+        res_log.write('\n* Description: ' + iptcmd_in.description)
+        res_log.write('\n* Headline: ' + iptcmd_in.headline)
+        res_log.write('\n* Instructions: ' + iptcmd_in.instructions)
+        res_log.write('\n* Intellectual Genre: ' + iptcmd_in.intellectualGenre)
+        res_log.write('\n* Job Id: ' + iptcmd_in.jobid)
+        res_log.write('\n* Keywords (multiple): ' + ', '.join(iptcmd_in.keywords))
+        res_log.write('\n* Rights Usage Terms: ' + iptcmd_in.usageTerms)
+        res_log.write('\n* Scene Codes (multiple): ' + ', '.join(iptcmd_in.sceneCodes))
+        res_log.write('\n* Source: ' + iptcmd_in.source)
+        res_log.write('\n* Subject Codes (multiple): ' + ', '.join(iptcmd_in.subjectCodes))
+        res_log.write('\n* Title: ' + iptcmd_in.title)
         res_log.write('\n***** IPTC Extension Schema')
-        res_log.write('\nLocation Created: ' + json.dumps(iptcmd_in.locationCreated.todict(), indent=2))
+        aboutCvTerm_list = []
+        for aboutCvTerm in iptcmd_in.aboutCvTerms:
+            aboutCvTerm_dict = aboutCvTerm.todict()
+            aboutCvTerm_list.append(aboutCvTerm_dict)
+        res_log.write('\n* About CV-Terms: ' + json.dumps(aboutCvTerm_list, indent=2))
+        res_log.write('\n* Location Created: ' + json.dumps(iptcmd_in.locationCreated.todict(), indent=2))
         locationsShown_list = []
         for locationShown in iptcmd_in.locationsShown:
             locationShown_dict = locationShown.todict()
             locationsShown_list.append(locationShown_dict)
-        res_log.write('\nLocations Shown: ' + json.dumps(locationsShown_list, indent=2))
+        res_log.write('\n* Locations Shown: ' + json.dumps(locationsShown_list, indent=2))
+        personsShown_list = []
+        for personShown in iptcmd_in.personsShown:
+            personShown_dict = personShown.todict()
+            personsShown_list.append(personShown_dict)
+        res_log.write('\n* Persons Shown: ' + json.dumps(personsShown_list, indent=2))
 
 # ************** MAIN